Revert "Merge branch 'develop' of https://github.com/Joseph-Lartey/N.A.C.K into frontend-validation"
This reverts commit 8fc5eeaf9c81ff244a15fb8fd5902aa53cc40e6f, reversing
changes made to f47cb4884e23b6ccf075c67916ee92080d20385e.
