+++ frontend/android/app/src/main/AndroidManifest.xml
@@ -3,7 +3,6 @@
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
 <uses-permission android:name="android.permission.USE_BIOMETRIC" />
-<uses-permission android:name="android.permission.INTERNET"/>
 <uses-permission android:name="android.permission.USE_FINGERPRINT" />
 
     <application
