+++ frontend/lib/main.dart
@@ -1,19 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:provider/provider.dart';
-import 'package:untitled3/providers/auth_provider.dart';
-import 'services/otp.dart'; // Import the OTP service
-import 'screens/WelcomeScreen.dart';
+import 'otp.dart'; // Import the OTP service
+import 'WelcomeScreen.dart';
 
 void main() {
   SystemChrome.setSystemUIOverlayStyle(const SystemUiOverlayStyle(
     statusBarColor: Colors.transparent,
   ));
   OTPService.configure(); // Configure the OTP service
-  runApp(ChangeNotifierProvider(
-    create: (context) => AuthProvider(),
-    child: const MyApp(),
-  ));
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
