+++ frontend/lib/otpPage.dart
@@ -1,26 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'package:untitled3/providers/auth_provider.dart';
-import '../services/otp.dart'; // Import the OTP service
+import 'otp.dart'; // Import the OTP service
 import 'profilesetup.dart'; // Import the InterestsPage
 
 class OtpPage extends StatefulWidget {
   final String email;
-  final String firstname;
-  final String lastname;
-  final String password;
-  final String confirmPassword;
-  final String dob;
 
-  const OtpPage(
-      {Key? key,
-      required this.email,
-      required this.firstname,
-      required this.lastname,
-      required this.password,
-      required this.confirmPassword,
-      required this.dob})
-      : super(key: key);
+  const OtpPage({Key? key, required this.email}) : super(key: key);
 
   @override
   State<OtpPage> createState() => _OtpPageState();
@@ -31,34 +16,16 @@ class _OtpPageState extends State<OtpPage> {
 
   bool _isButtonEnabled = false;
 
-  void _verifyOTP() async {
+  void _verifyOTP() {
     bool isValid = OTPService.verifyOTP(otpController.text);
-    final authProvider = Provider.of<AuthProvider>(context, listen: false);
-
     if (isValid) {
       ScaffoldMessenger.of(context).showSnackBar(
         const SnackBar(content: Text('OTP verified successfully!')),
       );
-
-      // make register request to api
-      await authProvider.register(
-        widget.firstname,
-        widget.lastname,
-        widget.firstname,
-        widget.email,
-        widget.password,
-        widget.dob,
+      Navigator.push(
+        context,
+        _createRoute(const ProfileSetupPage()), // Using custom route transition
       );
-
-      if (authProvider.registrationSuccess == true) {
-        print("registration finished");
-        // navigate to profile setup page
-        Navigator.push(
-          context,
-          _createRoute(
-              const ProfileSetupPage()), // Using custom route transition
-        );
-      }
     } else {
       ScaffoldMessenger.of(context).showSnackBar(
         const SnackBar(content: Text('Invalid OTP! Please try again.')),
