+++ frontend/lib/regScreen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../services/otp.dart'; // Import the OTP service
+import 'otp.dart'; // Import the OTP service
 import 'otpPage.dart'; // Import the OTP page
 import 'loginScreen.dart';
 import 'package:intl/intl.dart'; // Import intl package for date formatting
@@ -63,17 +63,13 @@ class _RegScreenState extends State<RegScreen> {
   }
 
   void _sendOTP() {
-    OTPService.sendOTP(emailController.text);
-    Navigator.push(
-      context,
-      _createRoute(OtpPage(
-          email: emailController.text,
-          firstname: firstNameController.text,
-          lastname: lastNameController.text,
-          password: passwordController.text,
-          confirmPassword: confirmPasswordController.text,
-          dob: dobController.text)),
-    );
+    if (_formKey.currentState?.validate() ?? false) {
+      OTPService.sendOTP(emailController.text);
+      Navigator.push(
+        context,
+        _createRoute(OtpPage(email: emailController.text)),
+      );
+    }
   }
 
   String? _validateEmail(String? value) {
