+++ frontend/pubspec.yaml
@@ -40,14 +40,11 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  email_otp: ^3.0.2
+  email_otp: ^3.0.1
   intl: ^0.17.0
   shared_preferences: ^2.0.11
   local_auth: ^2.1.4
   curved_navigation_bar: ^1.0.3
-  http:
-  provider:
-  quickalert:
 
 dev_dependencies:
   flutter_test:
